+++ docs/conf.py
@@ -1,6 +1,3 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
 extensions = [
     'sphinx.ext.autodoc',
     'jaraco.packaging.sphinx',
@@ -12,7 +9,7 @@ html_theme = "furo"
 # Link dates and other references in the changelog
 extensions += ['rst.linker']
 link_files = {
-    '../CHANGES.rst': dict(
+    '../NEWS.rst': dict(
         using=dict(GH='https://github.com'),
         replace=[
             dict(
@@ -33,6 +30,7 @@ link_files = {
 
 # Be strict about any broken references
 nitpicky = True
+nitpick_ignore = []
 
 # Include Python intersphinx mapping to prevent failures
 # jaraco/skeleton#51
@@ -45,3 +43,14 @@ intersphinx_mapping = {
 autodoc_preserve_defaults = True
 
 extensions += ['jaraco.tidelift']
+
+# jaraco/jaraco.collections#11
+nitpick_ignore += [
+    ('py:class', 'v, remove specified key and return the corresponding value.'),
+    ('py:class', 'None.  Update D from dict/iterable E and F.'),
+    ('py:class', 'D[k] if k in D, else d.  d defaults to None.'),
+]
+
+nitpick_ignore += [
+    ('py:class', 're.Pattern'),
+]